# docs/swagger.yaml
definitions:
  model.Data:
    properties:
      key:
        type: string
      value:
        type: object
    type: object
  model.Error:
    properties:
      ErrorCode:
        type: string
      Errormessage:
        type: string
    type: object
  model.PaginationResponse:
    properties:
      totalPageCount:
        type: integer
    type: object
  model.Response:
    properties:
      data:
        items:
          $ref: '#/definitions/model.Data'
        type: array
      pagination:
        $ref: '#/definitions/model.PaginationResponse'
    type: object
info:
  contact: {}
paths:
  /api/cache-service/v1/data:
    get:
      parameters:
      - description: Page number
        in: query
        name: page_num
        required: true
        type: integer
      - description: Page size
        in: query
        name: page_size
        required: true
        type: integer
      responses:
        "200":
          description: fetches the data from cache
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: 'ErrorCode: ERR_BAD_REQUEST'
          schema:
            $ref: '#/definitions/model.Error'
        "500":
          description: 'ErrorCode: ERR_INTERNAL_SERVER'
          schema:
            $ref: '#/definitions/model.Error'
      summary: Fetch data from cache using pagination
      tags:
      - Cache service
    post:
      consumes:
      - application/json
      parameters:
      - description: Any valid json Data can be provided
        in: body
        name: Data
        required: true
        schema:
          type: object
      responses:
        "201":
          description: store data
          schema:
            type: ""
        "400":
          description: 'ErrorCode: ERR_BAD_REQUEST'
          schema:
            $ref: '#/definitions/model.Error'
        "500":
          description: 'ErrorCode: ERR_INTERNAL_SERVER'
          schema:
            $ref: '#/definitions/model.Error'
      summary: Store data in cache with persistence backup in DB
      tags:
      - Cache service
swagger: "2.0"